fix: 영양소 타입 int에서 double로 변경
각 영양소 업데이트 하는 부분 소수점 2자리까지로 제한

+++ src/main/java/com/f_log/flog/diet/domain/Diet.java
@@ -27,25 +27,25 @@ public class Diet extends BaseEntity {
     private UUID dietUuid;
 
     @Column(name = "total_carbohydrate")
-    private int totalCarbohydrate;
+    private double totalCarbohydrate;
 
     @Column(name = "total_protein")
-    private int totalProtein;
+    private double totalProtein;
 
     @Column(name = "total_fat")
-    private int totalFat;
+    private double totalFat;
 
     @Column(name = "total_sodium")
-    private int totalSodium;
+    private double totalSodium;
 
     @Column(name = "total_cholesterol")
-    private int totalCholesterol;
+    private double totalCholesterol;
 
     @Column(name = "total_sugars")
-    private int totalSugars;
+    private double totalSugars;
 
     @Column(name = "total_calories")
-    private int totalCalories;
+    private double totalCalories;
 
     @Column(name = "meal_date")
     private LocalDate mealDate;
@@ -68,13 +68,13 @@ public class Diet extends BaseEntity {
     @Builder
     public Diet(UUID dietUuid,
                 Member member,
-                int totalCarbohydrate,
-                int totalProtein,
-                int totalFat,
-                int totalSodium,
-                int totalCholesterol,
-                int totalSugars,
-                int totalCalories,
+                double totalCarbohydrate,
+                double totalProtein,
+                double totalFat,
+                double totalSodium,
+                double totalCholesterol,
+                double totalSugars,
+                double totalCalories,
                 MealType mealType,
                 LocalDate mealDate) {
         this.dietUuid = dietUuid;
@@ -94,31 +94,31 @@ public class Diet extends BaseEntity {
         this.s3Url = s3Url;
     }
 
-    public void updateTotalCarbohydrate(int totalCarbohydrate) {
+    public void updateTotalCarbohydrate(double totalCarbohydrate) {
         this.totalCarbohydrate = totalCarbohydrate;
     }
 
-    public void updateTotalProtein(int totalProtein) {
+    public void updateTotalProtein(double totalProtein) {
         this.totalProtein = totalProtein;
     }
 
-    public void updateTotalFat(int totalFat) {
+    public void updateTotalFat(double totalFat) {
         this.totalFat = totalFat;
     }
 
-    public void updateTotalSodium(int totalSodium) {
+    public void updateTotalSodium(double totalSodium) {
         this.totalSodium = totalSodium;
     }
 
-    public void updateTotalCholesterol(int totalCholesterol) {
+    public void updateTotalCholesterol(double totalCholesterol) {
         this.totalCholesterol = totalCholesterol;
     }
 
-    public void updateTotalSugars(int totalSugars) {
+    public void updateTotalSugars(double totalSugars) {
         this.totalSugars = totalSugars;
     }
 
-    public void updateTotalCalories(int totalCalories) {
+    public void updateTotalCalories(double totalCalories) {
         this.totalCalories = totalCalories;
     }
